Refact: adaptando o codigo para salvar o id do usuario no token de sessao e a partir dele todo o sistema faz o resto das operacoes

+++ entitys/loginUser.go
@@ -1,8 +1,8 @@
 package entitys
 
 type LoginUser struct {
-	Id       uint64 `gorm: primarykey; autoincrement`
-	Login    string `gorm: unique; json:"login"`
+	Id       int    `gorm:primarykey;autoincrement`
+	Login    string `gorm:unique; json:"login"`
 	Password string `json:"password"`
 }
